requirements.txt
ffmpeg-python
rich
google-generativeai
python-dotenv
scenedetect
opencv-python-headless
scikit-image
